+ (untested) dag_ignore_list option
New option allows specification of comma-separated list of dag_id regexes to *not* clear.

+++ db-cleanup/airflow-db-cleanup.py
@@ -1,10 +1,16 @@
 """
 A maintenance workflow that you can deploy into Airflow to periodically clean out the DagRun, TaskInstance, Log, XCom, Job DB and SlaMiss entries to avoid having too much data in your Airflow MetaStore.
 
-airflow trigger_dag --conf '{"maxDBEntryAgeInDays":30}' airflow-db-cleanup
+examples:
+----------
+airflow trigger_dag --conf '{"maxDBEntryAgeInDays":30,"dagIgnoreList":""}' airflow-db-cleanup
+airflow trigger_dag --conf '{"dagIgnoreList":"my_special_dag,my_other_special_dag"}' airflow-db-cleanup
+airflow trigger_dag --conf '{"maxDBEntryAgeInDays":7}' airflow-db-cleanup
 
 --conf options:
+---------------
     maxDBEntryAgeInDays:<INT> - Optional
+    dagIgnoreList:<String> - Optional
 
 """
 from airflow.models import DAG, DagRun, TaskInstance, Log, XCom, SlaMiss, DagModel, Variable
@@ -14,6 +20,7 @@ from airflow.operators.python_operator import PythonOperator
 from datetime import datetime, timedelta
 import os
 import logging
+import re
 
 import dateutil.parser
 from sqlalchemy import func, and_
@@ -31,6 +38,7 @@ SCHEDULE_INTERVAL = "@daily"            # How often to Run. @daily - Once a day 
 DAG_OWNER_NAME = "operations"           # Who is listed as the owner of this DAG in the Airflow Web Server
 ALERT_EMAIL_ADDRESSES = []              # List of email address to send email alerts to if this job fails
 DEFAULT_MAX_DB_ENTRY_AGE_IN_DAYS = int(Variable.get("max_db_entry_age_in_days", 30)) # Length to retain the log files if not already provided in the conf. If this is set to 30, the job will remove those files that are 30 days old or older.
+DEFAULT_DAG_IGNORE_LIST = str(Variable.get("dag_ignore_list", "")) # comma-separated list of dag_id regexes to *not* clear
 ENABLE_DELETE = True                    # Whether the job should delete the db entries or not. Included if you want to temporarily avoid deleting the db entries.
 DATABASE_OBJECTS = [                    # List of all the objects that will be deleted. Comment out the DB objects you want to skip.
     {"airflow_db_model": DagRun, "age_check_column": DagRun.execution_date, "dag_id": DagRun.dag_id, "keep_last_run": True},
@@ -70,10 +78,16 @@ def print_configuration_function(**context):
     max_db_entry_age_in_days = None
     if dag_run_conf:
         max_db_entry_age_in_days = dag_run_conf.get("maxDBEntryAgeInDays", None)
+        dag_ignore_list = dag_run_conf.get("dagIgnoreList", None)
     logging.info("maxDBEntryAgeInDays from dag_run.conf: " + str(dag_run_conf))
+    logging.info("dagIgnoreList from dag_run.conf: " + str(dag_ignore_list))
     if max_db_entry_age_in_days is None:
         logging.info("maxDBEntryAgeInDays conf variable isn't included. Using Default '" + str(DEFAULT_MAX_DB_ENTRY_AGE_IN_DAYS) + "'")
         max_db_entry_age_in_days = DEFAULT_MAX_DB_ENTRY_AGE_IN_DAYS
+    if dag_ignore_list is None:
+        logging.info("dagIgnoreList conf variable isn't included. Using Default '" + str(DEFAULT_DAG_IGNORE_LIST) + "'")
+        dag_ignore_list = DEFAULT_DAG_IGNORE_LIST
+
     max_date = now() + timedelta(-max_db_entry_age_in_days)
     logging.info("Finished Loading Configurations")
     logging.info("")
@@ -83,10 +97,12 @@ def print_configuration_function(**context):
     logging.info("max_date:                 " + str(max_date))
     logging.info("enable_delete:            " + str(ENABLE_DELETE))
     logging.info("session:                  " + str(session))
+    logging.info("dag_ignore_list:          " + str(dag_ignore_list))
     logging.info("")
 
-    logging.info("Setting max_execution_date to XCom for Downstream Processes")
+    logging.info("Setting max_execution_date and dag_ignore_list to XCom for Downstream Processes")
     context["ti"].xcom_push(key="max_date", value=max_date.isoformat())
+    context["ti"].xcom_push(key="dag_ignore_list", value=dag_ignore_list.isoformat())
 
 print_configuration = PythonOperator(
     task_id='print_configuration',
@@ -100,6 +116,10 @@ def cleanup_function(**context):
     logging.info("Retrieving max_execution_date from XCom")
     max_date = context["ti"].xcom_pull(task_ids=print_configuration.task_id, key="max_date")
     max_date = dateutil.parser.parse(max_date) # stored as iso8601 str in xcom
+    
+    logging.info("Retrieving dag_ignore_list from XCom")
+    dag_ignore_list = context["ti"].xcom_pull(task_ids=print_configuration.task_id, key="dag_ignore_list")
+    dag_ignore_list = [s.strip() for s in dag_ignore_list.split(',')] # transform string from xcom to list
 
     airflow_db_model = context["params"].get("airflow_db_model")
     age_check_column = context["params"].get("age_check_column")
@@ -114,38 +134,42 @@ def cleanup_function(**context):
     logging.info("age_check_column:         " + str(age_check_column))
     logging.info("keep_last_run:            " + str(keep_last_run))
     logging.info("dag_id:                   " + str(dag_id))
+    logging.info("dag_ignore_list:          " + str(dag_ignore_list))
     logging.info("")
+    
+    if any([re.compile(r).match(dag_id) for r in dag_ignore_list]):
+        logger.info("dag_id found in dag_ignore_list; skipping cleanup for dag_id " + str(str(dag_id))
+    else: 
+        logging.info("Running Cleanup Process...")
+        query = session.query(airflow_db_model).options(load_only(age_check_column))
+        if keep_last_run:
+            # workaround for MySQL "table specified twice" issue
+            # https://github.com/teamclairvoyant/airflow-maintenance-dags/issues/41
+            sub_query = session.query(func.max(age_check_column)).group_by(dag_id).from_self()
+            query = query.filter(
+                age_check_column.notin_(sub_query), 
+                and_(age_check_column <= max_date)
+            )
+        else:
+            query = query.filter(age_check_column <= max_date,)
 
-    logging.info("Running Cleanup Process...")
-    query = session.query(airflow_db_model).options(load_only(age_check_column))
-    if keep_last_run:
-        # workaround for MySQL "table specified twice" issue
-        # https://github.com/teamclairvoyant/airflow-maintenance-dags/issues/41
-        sub_query = session.query(func.max(age_check_column)).group_by(dag_id).from_self()
-        query = query.filter(
-            age_check_column.notin_(sub_query), 
-            and_(age_check_column <= max_date)
-        )
-    else:
-        query = query.filter(age_check_column <= max_date,)
+        entries_to_delete = query.all()
 
-    entries_to_delete = query.all()
+        logging.info("Query : " +  str(query))
+        logging.info("Process will be Deleting the following " + str(airflow_db_model.__name__) + "(s):")
+        for entry in entries_to_delete:
+            logging.info("\tEntry: " + str(entry) + ", Date: " + str(entry.__dict__[str(age_check_column).split(".")[1]]))
 
-    logging.info("Query : " +  str(query))
-    logging.info("Process will be Deleting the following " + str(airflow_db_model.__name__) + "(s):")
-    for entry in entries_to_delete:
-        logging.info("\tEntry: " + str(entry) + ", Date: " + str(entry.__dict__[str(age_check_column).split(".")[1]]))
+        logging.info("Process will be Deleting " + str(len(entries_to_delete)) + " " + str(airflow_db_model.__name__) + "(s)")
 
-    logging.info("Process will be Deleting " + str(len(entries_to_delete)) + " " + str(airflow_db_model.__name__) + "(s)")
-
-    if ENABLE_DELETE:
-        logging.info("Performing Delete...")
-        #using bulk delete
-        query.delete(synchronize_session=False)
-        session.commit()
-        logging.info("Finished Performing Delete")
-    else:
-        logging.warn("You're opted to skip deleting the db entries!!!")
+        if ENABLE_DELETE:
+            logging.info("Performing Delete...")
+            #using bulk delete
+            query.delete(synchronize_session=False)
+            session.commit()
+            logging.info("Finished Performing Delete")
+        else:
+            logging.warn("You're opted to skip deleting the db entries!!!")
 
     logging.info("Finished Running Cleanup Process")
